requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.0.0
pandas>=2.0.0
typing-extensions>=4.0.0 